requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
websockets==12.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
psutil==5.9.6
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
asyncio-mqtt==0.16.1
pyaudio==0.2.11
aiortc==1.6.0
